pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-wannier90-workflows"
dynamic = ["version"] # read from aiida_wannier90_workflows/__init__.py
description = "Advanced AiiDA workflows for Wannier90"
authors = [{name = 'The AiiDA team', email = 'developers@aiida.net'}]
readme = "README.md"
license = { file = "LICENSE.txt" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: AiiDA",
    "Environment :: Plugins",
    "License :: Free To Use But Restricted",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering :: Physics",
    "Intended Audience :: Science/Research"
]
keywords = ["aiida", "plugin", "wannier90"]
requires-python = ">=3.9"
dependencies = [
    "aiida-core>=2.0,<3",
    "aiida-pseudo>=0.6",
    "aiida-quantumespresso>=4.4",
    "aiida-wannier90>=2.2",
    "click>=8.0",
    "colorama"
]

[project.urls]
Source = "https://github.com/aiidateam/aiida-wannier90-workflows"
Documentation = "https://aiida-wannier90-workflows.readthedocs.io"

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-cov",
    "pytest-regressions",
    "pgtest"
]
pre-commit = ["pre-commit", "pylint>=3"]
docs = [
    "aiida-core~=2.5",
    "docutils",
    "sphinx",
    "sphinx-copybutton",
    "sphinx-book-theme",
    "sphinx-click",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe"
]
dev = [
    "typer[all]",
    "upf_tools~=0.1",
    "rich",
    "typer",
    "upf-tools"
]
analysis = ["pandas", "tables", "scikit-learn"]

[project.scripts]
"aiida-wannier90-workflows" = "aiida_wannier90_workflows.cli:cmd_root"

[project.entry-points."aiida.calculations"]
"wannier90_workflows.stash_to_remote" = "aiida_wannier90_workflows.calculations.stash_to_remote:stash_to_remote"
"wannier90_workflows.split" = "aiida_wannier90_workflows.calculations.split:Wannier90SplitCalculation"

[project.entry-points."aiida.parsers"]
"wannier90_workflows.split" = "aiida_wannier90_workflows.parsers.split:Wannier90SplitParser"

[project.entry-points."aiida.workflows"]
"wannier90_workflows.base.wannier90" = "aiida_wannier90_workflows.workflows.base.wannier90:Wannier90BaseWorkChain"
"wannier90_workflows.base.open_grid" = "aiida_wannier90_workflows.workflows.base.open_grid:OpenGridBaseWorkChain"
"wannier90_workflows.base.pw2wannier90" = "aiida_wannier90_workflows.workflows.base.pw2wannier90:Pw2wannier90BaseWorkChain"
"wannier90_workflows.base.projwfc" = "aiida_wannier90_workflows.workflows.base.projwfc:ProjwfcBaseWorkChain"
"wannier90_workflows.wannier90" = "aiida_wannier90_workflows.workflows.wannier90:Wannier90WorkChain"
"wannier90_workflows.open_grid" = "aiida_wannier90_workflows.workflows.open_grid:Wannier90OpenGridWorkChain"
"wannier90_workflows.bands" = "aiida_wannier90_workflows.workflows.bands:Wannier90BandsWorkChain"
"wannier90_workflows.optimize" = "aiida_wannier90_workflows.workflows.optimize:Wannier90OptimizeWorkChain"
"wannier90_workflows.projwfcbands" = "aiida_wannier90_workflows.workflows.projwfcbands:ProjwfcBandsWorkChain"
"wannier90_workflows.split" = "aiida_wannier90_workflows.workflows.split:Wannier90SplitWorkChain"

[tool.flit.module]
name = "aiida_wannier90_workflows"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.tool-check]
generated-members = [
    "self.exit",
    "self.exit_codes.*",
    "Wannier90Calculation.exit_codes.*",
]

[tool.pylint.messages_control]
disable = [
    "too-many-ancestors",
    "invalid-name",
    "duplicate-code",
    "import-outside-toplevel",
    # 'bad-continuation',
    # 'locally-disabled',
    # 'logging-format-interpolation',
    # 'inconsistent-return-statements',
    # 'no-else-raise',
    'too-many-arguments',
    'too-many-branches',
    'too-many-locals',
    'too-many-statements',
    'protected-access',
    'no-member',
    # 'unspecified-encoding',

]

[tool.pylint.basic]
good-names = [
    'i',
    'j',
    'k',
    'l',
    'm',
    'n',
    'x',
    'y',
    'z',
    'mu',
    'qb',
    'df',
    'ax',
    'pw',
    'pk',
    'qe',
    'wannier',
    'SsspFamily',
    'StructureData',
    'UpfData',
    'HpCalculation',
    'PwCalculation',
    'SsspFamily',
    'PseudoDojoFamily',
    'CutoffsPseudoPotentialFamily',
    'Wannier90BandsWorkChain',
    'Wannier90OptimizeWorkChain',
]

[tool.pytest.ini_options]
minversion = '6.0'
testpaths = ['tests']
# python_files = "test_*.py example_*.py"
filterwarnings = [
    'ignore::DeprecationWarning:frozendict:',
    'ignore::DeprecationWarning:pkg_resources:',
    'ignore::DeprecationWarning:reentry:',
    'ignore::DeprecationWarning:sqlalchemy_utils:',
    "ignore::DeprecationWarning:aiida:",
    "ignore::DeprecationWarning:plumpy:",
    "ignore::DeprecationWarning:django:",
    "ignore::DeprecationWarning:yaml:",
]

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source = ["aiida_wannier90_workflows"]

[tool.isort]
# Configuration of [isort](https://isort.readthedocs.io)
profile = "black"
force_sort_within_sections = true
sections = [
    'FUTURE',
    'STDLIB',
    'THIRDPARTY',
    'AIIDA',
    'AIIDA_QUANTUMESPRESSO',
    'AIIDA_WANNIER90',
    'FIRSTPARTY',
    'LOCALFOLDER'
]
known_aiida = ['aiida']
known_aiida_quantumespresso = ['aiida_quantumespresso']
known_aiida_wannier90 = ['aiida_wannier90']

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39

[testenv]
usedevelop=True

[testenv:py{39,310,311,312}]
description = Run the test suite against a python version
extras = testing
commands = pytest {posargs}

[testenv:pre-commit]
description = Run the pre-commit checks
extras = pre-commit
commands = pre-commit run {posargs}

[testenv:docs]
description = Build the documentation
extras = docs
commands = sphinx-build -nW --keep-going -b html {posargs} docs/source docs/build/html
commands_post = echo "open file://{toxinidir}/docs/build/html/index.html"
"""
